#11 :hammer: 프래그먼트 관리 코드로 수정함
